Don't fail to GC a package if it didn't create a directory
Otherwise we get stuck forever with this package that cannot be GC'd.

+++ src/package.rs
@@ -543,7 +543,9 @@ impl InstalledPackage {
     #[instrument]
     pub async fn delete(&self) -> Result<()> {
         let dir = self.directory();
-        remove_dir_all(dir).await?;
+        if dir.try_exists()? {
+            remove_dir_all(dir).await?;
+        }
         Ok(())
     }
 }
